melhorias no app
